Refactor the validator
