+++ src/App/Bundle/MainBundle/Resources/config/Submission/services.yml
@@ -19,7 +19,7 @@ services:
     app_main.submission.validator:
         class: App\Bundle\MainBundle\Infrastructure\Submission\Validator
         arguments:
-            - @app_main.post.factory
             - @app_main.post.writer
-            - @app_main.submission.entity_converter
             - @app_main.submission.writer
+            - @app_main.post.resolver
+            - @app_main.clock
