+++ src/App/Domain/Clock/ClockInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Domain\Clock;
+
+/**
+ * A clock used to retrieve datetimes
+ */
+interface ClockInterface
+{
+    /**
+     * Retrieve the current time
+     *
+     * @return \DateTime
+     */
+    public function now();
+}
